tounriquet_project/userRole/serializers.py
from rest_framework import serializers
from userRole.models import UserRole

class UserRoleSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserRole
        fields = '__all__'
